Restructuring package for GCloud Func
